Normalise upload path to Unix/URI form on Windows
This is in an effort to address failures to download on Google Cloud due to the mix of path separators being unsupported. At this point, this will also alter the path shown to the user in the console output and Buildkite UI, which may not be desirable. Further discussion is needed.

Also, I expect the tests to fail, but I don't have a Windows Golang setup handy.